Pengumpulan Tugas 1

+++ Tugas 1/header/Stock.hpp
@@ -8,7 +8,7 @@ class Stock {
     private:
         // attributes
         string code;
-        int price;
+        unsigned long price;
         int quantity;
         Company company;
